requirements.txt
Appium-Python-Client==2.11.1
pytest==8.2.0
pytest-xdist==3.6.1
selenium==4.11.2